pyproject.toml
[build-system]
requires = ['setuptools >= 61.0.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'egi2bids'
version = '0.1.0'
description = 'BIDS conversion of EGI datasets (.mff)'
readme = 'README.md'
license = {file = 'LICENSE'}
requires-python = '>=3.8'
authors = [
    {name = 'Gwenael Birot', email = 'gwenael.birot@fcbg.ch'},
    {name = 'Mathieu Scheltienne', email = 'mathieu.scheltienne@fcbg.ch'},
]
maintainers = [
    {name = 'Gwenael Birot', email = 'gwenael.birot@fcbg.ch'},
    {name = 'Mathieu Scheltienne', email = 'mathieu.scheltienne@fcbg.ch'},
]
keywords = [
    'python',
    'bids',
    'eeg',
    'egi',
]
classifiers = [
    'Operating System :: Microsoft :: Windows',
    'Operating System :: Unix',
    'Operating System :: MacOS',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Natural Language :: English',
    'License :: OSI Approved :: MIT License',
]
dependencies = [
    'mne>=1.2',
    'mne-bids>=0.11',
    'numpy>=1.21',
    'pybv>=0.7.3',
    'psutil',
    'PyQt6',
]

[project.optional-dependencies]
build = [
    'build',
    'twine',
]
style = [
    'bibclean',
    'black',
    'codespell',
    'isort',
    'flake8',
    'pydocstyle[toml]',
]
test = [
    'pytest',
    'pytest-cov',
    'pybids',
]
all = [
    'egi2bids[style]',
    'egi2bids[test]',
    'egi2bids[build]',
]
full = [
    'egi2bids[all]',
]

[project.urls]
homepage = 'https://github.com/fcbg-hnp/egi2bids'
documentation = 'https://github.com/fcbg-hnp/egi2bids'
source = 'https://github.com/fcbg-hnp/egi2bids'
tracker = 'https://github.com/fcbg-hnp/egi2bids/issues'

[project.scripts]
egi2bids-sys_info = 'egi2bids.commands.sys_info:run'
egi2bids = 'egi2bids.commands.egi2bids:run'
egi2bids-gui = 'egi2bids.commands.gui:run'

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ['egi2bids*']
exclude = ['egi2bids*tests']

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
(
      __pycache__
    | \.github
    | setup.py
)
'''

[tool.isort]
profile = 'black'
multi_line_output = 3
line_length = 79
py_version = 38
extend_skip_glob = [
    'setup.py',
]

[tool.pydocstyle]
convention = 'numpy'
ignore-decorators = '(copy_doc|property|.*setter|.*getter|pyqtSlot|Slot)'
match = '^(?!setup|__init__|test_).*\.py'
match-dir = '^egi2bids.*'
add_ignore = 'D100,D104,D107'

[tool.pytest.ini_options]
minversion = '6.0'
addopts = '--durations 20 --junit-xml=junit-results.xml --verbose'
filterwarnings = []

[tool.coverage.run]
branch = true
cover_pylib = false
omit = [
    '**/__init__.py',
    '**/egi2bids/_version.py',
    '**/egi2bids/commands/*',
    '**/tests/**',
]

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'if __name__ == .__main__.:',
]
precision = 2
